Old/old_requirements.txt
langchain-community>=0.3,<0.4
langchain>=0.3,<0.4
langchain-ollama>=0.2,<0.3
langchain-core>=0.3,<0.4
chromadb
pypdf
marker-pdf